Added tests, fixed critical Bug, refactored
I added tests for Zero Sized Types (zst) because some interactions differ based
on whether the containing element (item inside sector) is a Zero Sized type
or not. Found a critical bug that would have caused undefined
behaviour in the _remove() function. A bit of refactoring for the
shrink/grow functions.