100% Test coverage for the Normal state (I think)
So I added the last tests for the draining functionality.
I also created a new struct for testing purposes that looks if the
Elements in the Sector get corretly dropped. This is only getting used
by the drain tests right now, but I should extend this to all tests in
the future, even though I am unsure if this is necessary, because if the
drain tests work every test regarding the items drop should work (in
theory).